requirements.txt
psycopg2-binary==2.9.3
django==4.0.5
python-dotenv==0.20
django-split-settings==1.1
django-debug-toolbar==3.4.0
pydantic==1.9.1
uwsgi==2.0.20
django-extensions==3.1.5
django-cors-headers==3.12.0
make==0.1.6.post2
redis==4.3.2
elasticsearch==8.2.2
celery==5.2.7